Moving code to setup core/kernel component ctypes from pyomo.environ to pyomo.kernel. This reduces the potential impact of merging this branch
