+++ pyomo/core/tests/unit/test_component_suffix.py
@@ -3,7 +3,7 @@ import collections
 import pickle
 
 import pyutilib.th as unittest
-import pyomo.core.kernel
+import pyomo.kernel
 from pyomo.core.tests.unit.test_component_dict import \
     _TestComponentDictBase
 from pyomo.core.tests.unit.test_component_list import \
